+++ processor/src/main/java/com/pwinckles/jdbcgen/processor/EntityAnalyzer.java
@@ -8,6 +8,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
 import java.util.function.Function;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 import javax.annotation.processing.ProcessingEnvironment;
 import javax.lang.model.element.ElementKind;
@@ -22,6 +23,9 @@ import javax.lang.model.type.TypeKind;
  * Analyzes an entity class to extract the metadata necessary to generate a {@link JdbcGenDb} instance.
  */
 public class EntityAnalyzer {
+
+    private static final Pattern UNESCAPED_QUOTE_PATTERN = Pattern.compile("(?<!\\\\)\"");
+    private static final String ESCAPED_QUOTE = "\\\\\"";
 
     private final ProcessingEnvironment processingEnv;
 
@@ -56,7 +60,7 @@ public class EntityAnalyzer {
         }
 
         var tableAnnotation = getAndValidateTableAnnotation(entity);
-        entitySpecBuilder.withTableName(tableAnnotation.name());
+        entitySpecBuilder.withTableName(escapeQuotes(tableAnnotation.name()));
 
         var fields = getAndValidateFields(entity);
 
@@ -198,7 +202,7 @@ public class EntityAnalyzer {
             var fieldIsPrivate = field.getModifiers().contains(Modifier.PRIVATE);
             var columnAnnotation = field.getAnnotationsByType(JdbcGenColumn.class)[0];
             var builder = FieldSpec.builder()
-                    .withColumnName(columnAnnotation.name())
+                    .withColumnName(escapeQuotes(columnAnnotation.name()))
                     .withIdentity(columnAnnotation.identity())
                     .withFieldElement(field);
 
@@ -245,5 +249,9 @@ public class EntityAnalyzer {
         return BeanUtil.getterName(
                 field.getSimpleName().toString(),
                 processingEnv.getTypeUtils().getPrimitiveType(TypeKind.BOOLEAN).equals(field.asType()));
+    }
+
+    private String escapeQuotes(String value) {
+        return UNESCAPED_QUOTE_PATTERN.matcher(value).replaceAll(ESCAPED_QUOTE);
     }
 }